Moved math tests to test_math.rs
